Remove legacy tool modules
