+++ agent/tools/check_website_tool.py
@@ -1,7 +1,16 @@
 from pydantic import BaseModel, Field
 from langchain_core.tools import tool
 
-from .network_tools import check_website
+import requests
+
+
+def check_website(url: str) -> str:
+    """Return HTTP status code for a website or an error message."""
+    try:
+        resp = requests.head(url, timeout=5)
+        return f"{resp.status_code}"
+    except Exception as e:
+        return f"Error: {e}"
 
 class CheckWebsiteInput(BaseModel):
     url: str = Field(..., description="Website URL")
