+++ tests/test_github_tool.py
@@ -1,10 +1,10 @@
 import unittest
 from unittest.mock import patch
-from agent.tools.github_tool import get_repo_info
+from agent.tools.get_repo_info_tool import get_repo_info
 
 
 class GitHubToolTest(unittest.TestCase):
-    @patch('agent.tools.github_tool.requests.get')
+    @patch('agent.tools.get_repo_info_tool.requests.get')
     def test_get_repo_info_success(self, mock_get):
         mock_get.return_value.raise_for_status.return_value = None
         mock_get.return_value.json.return_value = {
@@ -16,7 +16,7 @@ class GitHubToolTest(unittest.TestCase):
         self.assertIn('octocat/Hello-World', result)
         mock_get.assert_called_once()
 
-    @patch('agent.tools.github_tool.requests.get', side_effect=Exception('fail'))
+    @patch('agent.tools.get_repo_info_tool.requests.get', side_effect=Exception('fail'))
     def test_get_repo_info_error(self, mock_get):
         result = get_repo_info('octocat/Hello-World')
         self.assertIn('Failed to fetch', result)
