+++ tests/test_pdf_tools.py
@@ -3,12 +3,9 @@ import sys
 import tempfile
 import unittest
 from pathlib import Path
-from agent.tools.pdf_tools import (
-    extract_text_from_pdf,
-    summarize_text,
-    summarize_pdf,
-    count_words_in_file,
-)
+from agent.tools.extract_text_from_pdf_tool import extract_text_from_pdf
+from agent.tools.summarize_pdf_tool import summarize_pdf, summarize_text
+from agent.tools.count_words_in_file_tool import count_words_in_file
 
 sys.path.insert(0, str(Path(__file__).resolve().parent.parent))
